chore: break out depracted classes, make new classes abstract
